Create BasicWebScraper.py
This is just a "work in progress" web scraper I've been playing with as a read more about requests, lxml, and beautiful soup. I've learned a significant amount about regular expressions since the creation of this code and I intend on implementing more functionality and putting this code to use for data science related projects.